requirements.txt
tqdm==4.62.3
matplotlib==3.4.3
ipykernel==6.25.2
ipywidgets==8.1.1
matplotlib==3.7.1
scipy==1.10.1
diffusers==0.17.1
Pillow==9.5.0